Provide means to start a MessagingConfigurer
Since we decide to delegate to lower level Configurer instances, we need
 a means to start the high level implementations as well. To that end, a
 specific interface implementing the Configurer and providing the start
 operation is beneficial. This interface should be implemented by the
 RootConfigurer and the MessagingConfigurer to allow either to be
 started. The MessagingConfigurer can use the root(Consumer) operation
 to retrieve the delegate RootConfigurer, to invoke start on it for this
 purpose.

#3067
