+++ messaging/src/main/java/org/axonframework/configuration/RootConfiguration.java
@@ -18,8 +18,7 @@ package org.axonframework.configuration;
 
 /**
  * The root configuration of any Axon Framework application. Provides a means to {@link #start()} and
- * {@link #shutdown() stop} the application, besides containing all {@link Component components} and
- * {@link Module modules}.
+ * {@link #shutdown() stop} the application, besides containing all {@link Component components}.
  *
  * @author Allard Buijze
  * @author Steven van Beelen
@@ -28,12 +27,18 @@ package org.axonframework.configuration;
 public interface RootConfiguration extends LifecycleSupportingConfiguration {
 
     /**
-     * Starts this configuration. All components defined in this {@code RootConfiguration} will be started.
+     * All components defined in this {@code RootConfiguration} will be started.
+     * <p>
+     * Starting a {@code RootConfiguration} typically results in the invocation of all
+     * {@link #onStart(int, LifecycleHandler) registered start handlers}.
      */
     void start();
 
     /**
      * Shuts down the components defined in this {@code RootConfiguration}.
+     * <p>
+     * Shutting down a {@code RootConfiguration} typically results in the invocation of all
+     * {@link #onShutdown(int, LifecycleHandler) registered shutdown handlers}.
      */
     void shutdown();
 }
